requirements.txt

pypdf2 == 3.0.1
streamlit == 1.35.0
google-generativeai == 0.6.0